moved IDs to Long for Lab 10

+++ Lab-10-petclinic-solution/src/main/java/com/petclinic/core/OwnerController.java
@@ -13,7 +13,7 @@ public class OwnerController {
     }
 
     @RequestMapping("/owner/{id}")
-    public Owner findById(@PathVariable int id) {
+    public Owner findById(@PathVariable Long id) {
         return this.ownerService.findById(id);
     }
